Wrap inline integration components inside useCallback
We create a component named "TaxiSection" to prevent code duplication.

Everytime the view is called (ie. every second for the operator page
since useSWR refresh interval is set to 1000), a new instance of the
callback was generated so the whole section was replaced even if nothing
has changed.

When the user started to enter a new vehicle location, the input was
automatically replaced with the original value after max 1s.