components: DataGrid had a breaking API change
Oddly enough, it still worked in some cases.

This change was adverted deep in one of their 37 alpha releases, really
nice of them not to announce all of the breaking changes in the stable
v4.0 release from the previous stable version.

hideFooterRowCount is part of the commercial version.

Bump bugfix version while I'm testing this.