integration: address popup to select coordinates
Not what I wanted to do initially on the client page, I had more what I
did on the operator page, but it wasn't possible. Or it would require
rewriting half of the code.

It might happen though, when I address the other issue of not reusing
the selected coordinates in the hail form.